refactor(config_routes): extract security.py + governance.py from _factory.py (4/8)
Schritt 4 von 8 im Split-Plan
(`docs/superpowers/plans/2026-04-29-config-routes-split.md`).

What moved
- `_register_security_routes()` (~975 LOC, groesster Helper im Paket) ->
  `cognithor.channels.config_routes.security`. SecurityScanner, Compliance,
  Decision-Log, Remediation, Compliance-Export, Security-Pipeline /
  Ecosystem-Policy, Metrics, Incident-Tracker, Security-Team,
  Posture-Scorer, Security-Gate, Continuous-RedTeam, Code-Auditor.
- `_register_governance_routes()` (~218 LOC) ->
  `cognithor.channels.config_routes.governance`. Reputation-Engine,
  Recall-Manager, Abuse-Reporter, Governance-Policy, Interop,
  Economic-Governor, Hub, Impact-Assessor, Ecosystem-Controller.

`_factory.py` schrumpft von 4 794 -> 3 603 Zeilen (-25 %).

Per-Sub-Module-Imports
- `security.py`: `Request`, `HTTPException`, `Any`, `log`. F401-clean.
- `governance.py`: `HTTPException`, `Any`, `log`. F401-clean.

Aufruf-Reihenfolge in `create_config_routes()` exakt wie zuvor; Drift-Test
gruen (303/303 Endpoints).

Verification
- `pytest test_route_inventory_unchanged` -> passed.
- Konsumenten-Sweep ueber 11 Test-Files -> 396 passed.
- `ruff check` -> All checks passed.
- `ruff format --check` -> 8 files already formatted.
